refactor: moved hypha to one file
